More code

+++ include/game_core/tank_ai.h
@@ -58,9 +58,9 @@ namespace AI{
 	\param target_id The id of the target tank whose distance to check
 	\return The distance between the 2 tanks
 	*/
-	CUDA_EXPORT uint32_t get_tank_dist(BasicTank::TankCollection* tc,
-									   tank_id my_id,
-									   tank_id target_id);
+	CUDA_EXPORT f32 get_tank_dist(BasicTank::TankCollection* tc,
+								  tank_id my_id,
+								  tank_id target_id);
 	
 	/*!
 	Get the distance of a bullet to the current tank
@@ -70,7 +70,7 @@ namespace AI{
 	\param The id of the bullet to query
 	\return The distance between the bullet and the tank
 	*/
-	CUDA_EXPORT uint32_t get_bullet_dist(BasicTank::TankCollection* tc,
+	CUDA_EXPORT f32 get_bullet_dist(BasicTank::TankCollection* tc,
 										 TankBullet::BulletCollection* bc,
 										 tank_id tid,
 										 bullet_id bid);
